requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# LLM APIs
openai>=1.0.0

# PDF Processing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
camelot-py[cv]>=0.11.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.0
XlsxWriter>=3.1.0

# Document Generation
python-docx>=1.1.0

# Embeddings and Vector Storage
sentence-transformers>=2.7.0
faiss-cpu>=1.8.0
chromadb>=0.5.0

# Data Validation
pydantic>=2.8.0
pydantic-settings>=2.4.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging and Monitoring
loguru>=0.7.0
structlog>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0

# Utilities
python-dateutil>=2.9.0
requests>=2.32.0
tenacity>=8.5.0

# Optional: Web scraping for RNS/news
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Development
black>=24.0.0
flake8>=7.0.0
mypy>=1.11.0
isort>=5.13.0
